requirements.txt
google-generativeai
faiss-cpu
sentence-transformers
torch
clip-anytorch
pillow
pytesseract
requests
tqdm
aiohttp
python-dotenv
streamlit